utils: Move SPDX-related tests to their own file
Follow the pattern to name the test file after the file that contains
the code to test.
